+++ src/main/java/erp/retirement/model/RetirementTaxDeferral.java
@@ -8,33 +8,33 @@ import java.util.Date;
 public class RetirementTaxDeferral {
 
 	// [DB 관리 항목]
-	private Long retirementTaxDeferralId; // 과세이연 내역 식별 번호 (PK)
-	private Long retirementCalculationId; // 퇴직급여 계산내역 식별 번호 (FK)
+	private int retirementTaxDeferralId; // 과세이연 내역 식별 번호 (PK)
+	private int retirementCalculationId; // 퇴직급여 계산내역 식별 번호 (FK)
 
 	// [퇴직연금계좌 입금정보]
 	private String bizName; // 퇴직연금사업자명
 	private String bizRegNo; // 사업자등록번호
 	private String accountNo; // 계좌번호
 	private Date depositDate; // 입금(이체)일
-	private Long depositAmt; // 계좌입금금액
+	private long depositAmt; // 계좌입금금액
 
 	public RetirementTaxDeferral() {
 	}
 
 	// Getter & Setter
-	public Long getRetirementTaxDeferralId() {
+	public int getRetirementTaxDeferralId() {
 		return retirementTaxDeferralId;
 	}
 
-	public void setRetirementTaxDeferralId(Long retirementTaxDeferralId) {
+	public void setRetirementTaxDeferralId(int retirementTaxDeferralId) {
 		this.retirementTaxDeferralId = retirementTaxDeferralId;
 	}
 
-	public Long getRetirementCalculationId() {
+	public int getRetirementCalculationId() {
 		return retirementCalculationId;
 	}
 
-	public void setRetirementCalculationId(Long retirementCalculationId) {
+	public void setRetirementCalculationId(int retirementCalculationId) {
 		this.retirementCalculationId = retirementCalculationId;
 	}
 
@@ -70,11 +70,11 @@ public class RetirementTaxDeferral {
 		this.depositDate = depositDate;
 	}
 
-	public Long getDepositAmt() {
+	public long getDepositAmt() {
 		return depositAmt;
 	}
 
-	public void setDepositAmt(Long depositAmt) {
+	public void setDepositAmt(long depositAmt) {
 		this.depositAmt = depositAmt;
 	}
 }
